patch: fixed npm start and ignored pack
