+++ .gitignore
@@ -36,3 +36,6 @@ yarn-error.log
 
 # API keys and secrets
 .env
+
+#pack file
+simply-express-generator-*.tgz
