+++ package.json
@@ -4,11 +4,11 @@
   "description": "A simple express generator utility for barebones Express/TS projects",
   "main": "dist/index.js",
   "bin": {
-    "simply-express-generator": "dist/index.js"
+    "simply-express-generator": "dist/src/index.js"
   },
   "scripts": {
     "build": "tsc",
-    "start": "node ./dist/index.js",
+    "start": "node ./dist/src/index.js",
     "dev": "ts-node -r tsconfig-paths/register ./src/index.ts",
     "prepublishOnly": "npm run build",
     "prepare": "npm run build",
@@ -68,6 +68,5 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/marcosppollastri/simply-express-generator"
-  },
-  "eslintIgnore": ["**/template/**"]
+  }
 }
